requirements.txt
# requirements.txt

# Core
flask>=3.0
flask-cors>=4.0
pymongo>=4.6
python-dotenv>=1.0
bcrypt>=4.1
PyJWT>=2.8
APScheduler>=3.10
requests>=2.31


# ML / Analytics
scikit-learn>=1.4
numpy>=1.26
joblib>=1.3

# NLP alias generation
spacy>=3.7
beautifulsoup4>=4.12
# After install run:  python -m spacy download en_core_web_sm

# QR Code
qrcode[pil]>=7.4
Pillow>=10.0
